backend/pyproject.toml
[project]
name = "chat-api"
version = "0.1.0"
description = "A simple chat thread API with Flask"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "uvicorn>=0.27.0",
    "asgiref>=3.7.0",
    "pyyaml>=6.0.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "psycopg[binary]>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
